+++ CHANGELOG.md
@@ -4,6 +4,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.3.0] - 2019-10-30
+### Added
+- [Pubsub] Full CloudEvents support
+- [Pubsub] Support for Pull Subscriptions
+- [Pubsub] Fine grained control over Subscription configurations
+- [Pubsub] Added Push support
+- [HTTP] Middleware Support and a built in Logger Middleware
+
+### Fixed
+- Correctly pass the event type
+
 ## [1.2.0] - 2019-08-22
 ### Added
 - Support for Pubsub publishing, based on CloudEvents